+++ DiscreteWalker/src/observables/Observables.h
@@ -47,7 +47,7 @@ double
 conditional_fpt_mean(const double s0,const double x, DiscreteWalker &walker, const long N,const bool target=1);
 
 double
-survival_probability(const double s0,const long n, DiscreteWalker &walker, const long N);
+survival_probability(const long s0,const long n, DiscreteWalker &walker, const long N);
 
 double
-splitting_probability(const double s0,const double x, DiscreteWalker &walker, const long N);
+splitting_probability(const long s0,const double x, DiscreteWalker &walker, const long N);
